# gdtracker-go-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GDTracker Go API
  version: 0.5.0
  description: |
    Go backend for GDTracker. Session auth and CSRF mirror the Spring API contract (see `docs/AUTH.md`).
    Routes mirror `gdtracker-api` for games, categories, tags, tasks, features, game configuration, archives, game exceptions,
    ingest plane (players, integration ping), game events (definitions + ingest), traces (heatmap), and feedback.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/games/{gameId}/tags:
    get:
      tags: [Tags]
      operationId: listTags
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        "401":
          description: Not authenticated
        "404":
          description: Game not found or not owned
    post:
      tags: [Tags]
      operationId: createTag
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpsertRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          description: Validation error
        "401":
          description: Not authenticated
        "404":
          description: Game not found or not owned
        "409":
          description: Duplicate tag name
  /api/games/{gameId}/tags/{id}:
    put:
      tags: [Tags]
      operationId: updateTag
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/TagId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpsertRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          description: Validation error
        "401":
          description: Not authenticated
        "404":
          description: Tag not found
        "409":
          description: Duplicate name or tag in use
    delete:
      tags: [Tags]
      operationId: deleteTag
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "404":
          description: Tag not found
        "409":
          description: Tag in use
...
  /api/games/{gameId}/game-players:
    post:
      tags: [IngestPlane]
      operationId: registerGamePlayer
      summary: Register a player for ingest (Bearer ingest token only)
      security: []
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [playerId]
                properties:
                  playerId: { type: string }
        "401": { description: Invalid token }
        "404": { description: Game not found }
  /api/games/{gameId}/integration:
    post:
      tags: [IngestPlane]
      operationId: postIntegrationPing
      summary: Record integration validation ping
      security: []
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationPingRequest"
      responses:
        "204": { description: No content }
        "400": { description: validation must be ok }
        "401": { description: Invalid token }
        "404": { description: Game not found }
  /api/games/{gameId}/integration/status:
    get:
      tags: [IngestPlane]
      operationId: getIntegrationStatus
      summary: Last successful integration ping time (session)
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationStatusResponse"
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
...
components:
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema: { type: string }
...
    TagId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ingest-token
  schemas:
...
    Tag:
      type: object
      required: [id, name, color]
      properties:
        id: { type: string }
        name: { type: string }
        color: { type: string }
        description: { type: string, nullable: true }
    TagUpsertRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        color: { type: string }
        description: { type: string, nullable: true }
...
    IntegrationPingRequest:
      type: object
      required: [validation]
      properties:
        validation:
          type: string
          enum: [ok]
    IntegrationStatusResponse:
      type: object
      properties:
        lastValidatedAt: { type: string, format: date-time, nullable: true }
...